+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v0.x.x
+- Added `instance` module
+
 ## v0.4.0
 - Added `appliance_settings` module
 - Added `appliance_maintenance_mode` module